+++ lib/gui/app.js
@@ -47,6 +47,7 @@ const app = angular.module('Etcher', [
   require('./models/settings'),
   require('./models/supported-formats'),
   require('./models/drives'),
+  require('./models/flash-state'),
 
   // Components
   require('./components/progress-button/progress-button'),
@@ -95,9 +96,9 @@ app.run((AnalyticsService, UpdateNotifierService, SelectionStateModel) => {
 
 });
 
-app.run((AnalyticsService, OSWindowProgressService, ImageWriterService) => {
+app.run((AnalyticsService, OSWindowProgressService, FlashStateModel) => {
   Store.subscribe(() => {
-    const flashState = ImageWriterService.getFlashState();
+    const flashState = FlashStateModel.getFlashState();
 
     // There is usually a short time period between the `isFlashing()`
     // property being set, and the flashing actually starting, which
@@ -106,7 +107,7 @@ app.run((AnalyticsService, OSWindowProgressService, ImageWriterService) => {
     //
     // We use the presence of `.eta` to determine that the actual
     // writing started.
-    if (!ImageWriterService.isFlashing() || !flashState.eta) {
+    if (!FlashStateModel.isFlashing() || !flashState.eta) {
       return;
     }
 
@@ -135,6 +136,7 @@ app.controller('AppController', function(
   $state,
   DriveScannerService,
   SelectionStateModel,
+  FlashStateModel,
   SettingsModel,
   SupportedFormatsModel,
   DrivesModel,
@@ -152,6 +154,7 @@ app.controller('AppController', function(
   this.selection = SelectionStateModel;
   this.drives = DrivesModel;
   this.writer = ImageWriterService;
+  this.state = FlashStateModel;
   this.settings = SettingsModel;
   this.tooltipModal = TooltipModalService;
 
@@ -174,7 +177,7 @@ app.controller('AppController', function(
   // This catches the case where the user enters
   // the settings screen when a flash finished
   // and goes back to the main screen with the back button.
-  if (!this.writer.isFlashing()) {
+  if (!FlashStateModel.isFlashing()) {
 
     this.selection.clear({
 
@@ -248,7 +251,7 @@ app.controller('AppController', function(
   };
 
   this.reselectImage = () => {
-    if (this.writer.isFlashing()) {
+    if (FlashStateModel.isFlashing()) {
       return;
     }
 
@@ -263,7 +266,7 @@ app.controller('AppController', function(
   };
 
   this.reselectDrive = () => {
-    if (this.writer.isFlashing()) {
+    if (FlashStateModel.isFlashing()) {
       return;
     }
 
@@ -276,12 +279,12 @@ app.controller('AppController', function(
       preserveImage: true
     });
 
-    this.writer.resetState();
+    FlashStateModel.resetState();
     AnalyticsService.logEvent('Restart after failure');
   };
 
   this.wasLastFlashSuccessful = () => {
-    const flashResults = this.writer.getFlashResults();
+    const flashResults = FlashStateModel.getFlashResults();
 
     if (_.get(flashResults, 'cancelled', false)) {
       return true;
@@ -292,7 +295,7 @@ app.controller('AppController', function(
 
   this.flash = (image, drive) => {
 
-    if (this.writer.isFlashing()) {
+    if (FlashStateModel.isFlashing()) {
       return;
     }
 
@@ -306,7 +309,7 @@ app.controller('AppController', function(
     });
 
     return this.writer.flash(image, drive).then(() => {
-      const results = ImageWriterService.getFlashResults();
+      const results = FlashStateModel.getFlashResults();
 
       if (results.cancelled) {
         return;
