+++ lib/gui/modules/image-writer.js
@@ -22,157 +22,15 @@
 
 const angular = require('angular');
 const _ = require('lodash');
-const Store = require('../models/store');
 const childWriter = require('../../src/child-writer');
 
 const MODULE_NAME = 'Etcher.Modules.ImageWriter';
 const imageWriter = angular.module(MODULE_NAME, [
-  require('../models/settings')
+  require('../models/settings'),
+  require('../models/flash-state')
 ]);
 
-imageWriter.service('ImageWriterService', function($q, $rootScope, SettingsModel) {
-
-  /**
-   * @summary Reset flash state
-   * @function
-   * @public
-   *
-   * @example
-   * ImageWriterService.resetState();
-   */
-  this.resetState = () => {
-    Store.dispatch({
-      type: Store.Actions.RESET_FLASH_STATE
-    });
-  };
-
-  /**
-   * @summary Check if currently flashing
-   * @function
-   * @private
-   *
-   * @returns {Boolean} whether is flashing or not
-   *
-   * @example
-   * if (ImageWriterService.isFlashing()) {
-   *   console.log('We\'re currently flashing');
-   * }
-   */
-  this.isFlashing = () => {
-    return Store.getState().toJS().isFlashing;
-  };
-
-  /**
-   * @summary Set the flashing flag
-   * @function
-   * @private
-   *
-   * @description
-   * This function is extracted for testing purposes.
-   *
-   * The flag is used to signify that we're going to
-   * start a flash process.
-   *
-   * @example
-   * ImageWriterService.setFlashingFlag();
-   */
-  this.setFlashingFlag = () => {
-    Store.dispatch({
-      type: Store.Actions.SET_FLASHING_FLAG
-    });
-  };
-
-  /**
-   * @summary Unset the flashing flag
-   * @function
-   * @private
-   *
-   * @description
-   * This function is extracted for testing purposes.
-   *
-   * The flag is used to signify that the write process ended.
-   *
-   * @param {Object} results - flash results
-   *
-   * @example
-   * ImageWriterService.unsetFlashingFlag({
-   *   passedValidation: true,
-   *   cancelled: false,
-   *   sourceChecksum: 'a1b45d'
-   * });
-   */
-  this.unsetFlashingFlag = (results) => {
-    Store.dispatch({
-      type: Store.Actions.UNSET_FLASHING_FLAG,
-      data: results
-    });
-  };
-
-  /**
-   * @summary Set the flashing state
-   * @function
-   * @private
-   *
-   * @description
-   * This function is extracted for testing purposes.
-   *
-   * @param {Object} state - flashing state
-   *
-   * @example
-   * ImageWriterService.setProgressState({
-   *   type: 'write',
-   *   percentage: 50,
-   *   eta: 15,
-   *   speed: 100000000000
-   * });
-   */
-  this.setProgressState = (state) => {
-    Store.dispatch({
-      type: Store.Actions.SET_FLASH_STATE,
-      data: {
-        type: state.type,
-        percentage: state.percentage,
-        eta: state.eta,
-
-        speed: _.attempt(() => {
-          if (_.isNumber(state.speed) && !_.isNaN(state.speed)) {
-
-            // Transform bytes to megabytes preserving only two decimal places
-            return Math.floor(state.speed / 1e+6 * 100) / 100;
-
-          }
-        })
-      }
-    });
-  };
-
-  /**
-   * @summary Get the flash results
-   * @function
-   * @public
-   *
-   * @returns {Object} flash results
-   *
-   * @example
-   * const results = ImageWriterService.getFlashResults();
-   */
-  this.getFlashResults = () => {
-    return Store.getState().toJS().flashResults;
-  };
-
-  /**
-   * @summary Get the current flash state
-   * @function
-   * @public
-   *
-   * @returns {Object} flash state
-   *
-   * @example
-   * const flashState = ImageWriterService.getFlashState();
-   */
-  this.getFlashState = () => {
-    return Store.getState().get('flashState').toJS();
-  };
+imageWriter.service('ImageWriterService', function($q, $rootScope, SettingsModel, FlashStateModel) {
 
   /**
    * @summary Perform write operation
@@ -228,11 +86,11 @@ imageWriter.service('ImageWriterService', function($q, $rootScope, SettingsModel
    * });
    */
   this.flash = (image, drive) => {
-    if (this.isFlashing()) {
+    if (FlashStateModel.isFlashing()) {
       return $q.reject(new Error('There is already a flash in progress'));
     }
 
-    this.setFlashingFlag();
+    FlashStateModel.setFlashingFlag();
 
     return this.performWrite(image, drive, (state) => {
 
@@ -241,7 +99,7 @@ imageWriter.service('ImageWriterService', function($q, $rootScope, SettingsModel
       // `.getFlashState()` will not return
       // the latest updated progress state.
       $rootScope.$apply(() => {
-        this.setProgressState(state);
+        FlashStateModel.setProgressState(state);
       });
 
     }).then((results) => {
@@ -253,9 +111,9 @@ imageWriter.service('ImageWriterService', function($q, $rootScope, SettingsModel
         passedValidation: false
       });
 
-      this.unsetFlashingFlag(results);
+      FlashStateModel.unsetFlashingFlag(results);
     }).catch((error) => {
-      this.unsetFlashingFlag({
+      FlashStateModel.unsetFlashingFlag({
         cancelled: false,
         passedValidation: false,
         errorCode: error.code
