+++ lib/gui/pages/finish/controllers/finish.js
@@ -16,7 +16,7 @@
 
 'use strict';
 
-module.exports = function($state, ImageWriterService, SelectionStateModel, AnalyticsService, SettingsModel) {
+module.exports = function($state, FlashStateModel, SelectionStateModel, AnalyticsService, SettingsModel) {
 
   /**
    * @summary Settings model
@@ -30,7 +30,7 @@ module.exports = function($state, ImageWriterService, SelectionStateModel, Analy
    * @type String
    * @public
    */
-  this.checksum = ImageWriterService.getFlashResults().sourceChecksum;
+  this.checksum = FlashStateModel.getFlashResults().sourceChecksum;
 
   /**
    * @summary Restart the flashing process
