+++ lib/gui/pages/finish/finish.js
@@ -31,7 +31,7 @@ const MODULE_NAME = 'Etcher.Pages.Finish';
 const FinishPage = angular.module(MODULE_NAME, [
   require('angular-ui-router'),
   require('../../modules/analytics'),
-  require('../../modules/image-writer'),
+  require('../../models/flash-state'),
   require('../../models/selection-state'),
   require('../../models/settings')
 ]);
